requirements.txt
galois==0.0.24
torchvision==0.10.1
matplotlib==3.5.1
